More default resolvers to the server
