+++ packages/lesswrong/server.ts
@@ -187,6 +187,7 @@ import './server/resolvers/dialogueChecksResolvers';
 import './server/resolvers/databaseSettingsResolvers';
 import './server/resolvers/reviewWinnerResolvers';
 import './server/resolvers/googleVertexResolvers';
+import './server/resolvers/defaultResolvers';
 
 import './server/intercomSetup';
 import './server/callbacks/intercomCallbacks';
