+++ packages/lesswrong/server/recombee/client.ts
@@ -12,7 +12,7 @@ import groupBy from 'lodash/groupBy';
 import uniq from 'lodash/uniq';
 import { recommendationsTabManuallyStickiedPostIdsSetting } from '../../lib/publicSettings';
 import { getParentTraceId, openPerfMetric, wrapWithPerfMetric } from '../perfMetrics';
-import { performQueryFromViewParameters } from '../../lib/vulcan-core/default_resolvers';
+import { performQueryFromViewParameters } from '../resolvers/defaultResolvers';
 import { captureException } from '@sentry/core';
 import { randomId } from '../../lib/random';
